KontoWorkerRegistration
A simple recording project based on MVC that I did in the internship.
You can simply register for new recruits within the project. You can search by place of residence from registered people, which is important for this project.
A very necessary project for census companies.

It is available in the files in DB.

Hope it will be useful for you :)